:recycle: Move run_version_command and run_undo_command into command:: submodules
